requirements.txt
Django==1.7
dj-static==0.0.6
django-angular==0.7.10
Markdown==2.5.2
MySQL-python==1.2.5
Pillow==2.7.0
PyJWT==0.4.1
argparse==1.2.1
django-countries==3.1.1
django-extensions==1.4.9
django-filter==0.9.1
django-suit==0.2.12
djangorestframework==3.0.3
oauthlib==0.7.2
pydot==1.0.2
pyparsing==1.5.7
python-openid==2.2.5
python-social-auth==0.2.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
static3==0.5.1
sorl-thumbnail==11.12.1b
wsgiref==0.1.2
